pulled out nav component
