+++ frontend/src/components/CustomerView/CustomerView.js
@@ -1,19 +1,16 @@
 import { IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
-import SettingsAccessibilityIcon from "@mui/icons-material/SettingsAccessibility";
 import CloseIcon from "@mui/icons-material/Close";
 import React, { useState } from "react";
 import { formatItemName } from "../../utils/formatItemName";
 import { useBasket } from "../common/BasketContext";
 import "./CustomerView.css";
-import { ReactComponent as ReveilleLogo } from "../../img/reveille_logo.svg";
-import { useFontSize } from "../../utils/FontSizeProvider";
-import classNames from 'classnames';
+import { navBar } from "../common/navBar";
 
 const CustomerView = ({ menuItems }) => {
     const [panel, setPanel] = useState(null);
     const [currType, setCurrType] = useState(null);
-    const { toggleFontSize, toggleIconScale, fontSizeMultiplier } = useFontSize();
+    
     const {
         basket,
         increaseItemQuantity,
@@ -213,105 +210,11 @@ const CustomerView = ({ menuItems }) => {
         </aside>
     );
 
-    const Accessibility = () => {
-        const [showAccessibilityPanel, setShowAccessibilityPanel] = useState(false);
-        const [screenReaderEnabled, setScreenReaderEnabled] = useState(false);
-        const [longPressBtnEnabled, setLongPressBtnEnabled] = useState(false);
-        const [highContrastEnabled, setHighContrastEnabled] = useState(false);
-        
-        return (
-            <>
-                <button
-                    aria-label="accessibility options"
-                    onClick={() =>
-                        setShowAccessibilityPanel(!showAccessibilityPanel)
-                    }
-                    className="accessibilityBtn"
-                >
-                    <SettingsAccessibilityIcon style={{fill: "white"}}/>
-                </button>
-                {showAccessibilityPanel && (
-                    <div className="accessibilityContainer">
-                        <button
-                            onClick={() => setShowAccessibilityPanel(!showAccessibilityPanel)}
-                            className="closeBtn"
-                        >
-                            <CloseIcon />
-                        </button>
-
-                        <span className="accessibilityTitle">
-                            Accessibility Options
-                        </span>
-
-                        <div className="accessibilityOptions">
-                            <button
-                                className={classNames('accessibilityOptionBtn', { 'accessibilityOptionBtnActive': screenReaderEnabled })}
-                                onClick={() =>setScreenReaderEnabled(!screenReaderEnabled)}
-                                variant="contained"
-                            >
-                                {screenReaderEnabled ? 'Disable Screen Reader' : 'Enable Screen Reader'}
-                            </button>
-
-                            <button
-                                className={classNames('accessibilityOptionBtn', { 'accessibilityOptionBtnActive': fontSizeMultiplier !== 1 })}
-                                onClick={ () => {
-                                    toggleFontSize()
-                                    toggleIconScale()
-                                }}
-                                variant="contained"
-                            >
-                                {fontSizeMultiplier !== 1 ? 'Disable Large Text' : 'Enable Large Text'}
-                            </button>
-
-                            <button
-                                className={classNames('accessibilityOptionBtn', { 'accessibilityOptionBtnActive': longPressBtnEnabled })}
-                                onClick={() =>setLongPressBtnEnabled(!longPressBtnEnabled)}
-                                variant="contained"
-                            >
-                                {longPressBtnEnabled ? 'Disable Long Press Buttons' : 'Enable Long Press Buttons'}
-                            </button>
-
-                            <button
-                                className={classNames('accessibilityOptionBtn', { 'accessibilityOptionBtnActive': highContrastEnabled })}
-                                onClick={() =>setHighContrastEnabled(!highContrastEnabled)}
-                                variant="contained"
-                            >
-                                {highContrastEnabled ? 'Disable High Contrast' : 'Enable High Contrast'}
-                            </button>
-                        </div>
-                    </div>
-                )}
-            </>
-        );
-    };
-
-    const navBar = () => {
-        return (
-            <nav className="navBar">
-                <div className="navSide">
-                    <Accessibility />
-                </div>
-                
-                <header className="navHeader">
-                    <ReveilleLogo className="reveilleLogo" />
-                    <h1 className="navTitle">
-                        Rev's American Grill
-                    </h1>
-                </header>
-                <div className="navSide">
-                    {/* TODO: add content */}
-                </div>
-            </nav>
-        )
-    };
-
     return (
         <main>
             {navBar()}
 
-            <div className="bodyPanel"
-                style={{ fontSize: `${fontSizeMultiplier}rem` }}
-            >   
+            <div className="bodyPanel">   
                 <aside className="typeMenu">
                     {typeButton("Burgers")}
                     {typeButton("Baskets")}
