+++ frontend/src/components/common/navBar.css
@@ -1,0 +1,99 @@
+.navBar {
+    width: 100%;
+    background-color: var(--primary-color);
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    min-height: 6vh;
+    justify-content: space-between;
+    padding: 0 20px;
+}
+
+.navSide {
+    position: static; /* Corrected to a valid position value, adjust as needed */
+}
+
+.navHeader {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-inline: 10px;
+}
+
+.reveilleLogo {
+    width: 50px;
+    height: 45px;
+    margin-right: 5px;
+}
+
+.navTitle {
+    color: #C2A061;
+    font-weight: bold;
+    font-size: 2rem;
+}
+
+.accessibilityBtn {
+    scale: 1.3;
+    border-radius: 50%;
+    padding: 11px 13px 9px; 
+    border: 2px solid var(--white); 
+    background-color: var(--accessibility-blue);
+    transform: scale(1);
+}
+
+.accessibilityCloseBtn {
+    scale: 0.9;
+    background-color: var(--close);
+    color: white;
+    float: left;
+}
+
+.accessibilityTitle {
+    padding: 10px;
+    font-weight: bold;
+    font-size: inherit;
+}
+
+.accessibilityIcon {
+    transform: none;
+}
+
+.accessibilityOptions {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 10px;
+}
+
+.accessibilityOptionBtn {
+    background-color: var(--neutral-medium);
+    color: black;
+    padding-block: 5px;
+    padding-inline: 10px;
+    margin-bottom: 10px;
+    width: 100%;
+    font-size: inherit;
+} 
+
+.accessibilityOptionBtnActive {
+    background-color: var(--neutral-dark);
+    color: white;
+}
+
+.accessibilityContainer {
+    position: fixed;
+    background: white;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    z-index: 2000;
+    top: 10%;
+    font-size: 30px;
+}
+
+.closeBtn {
+    scale: 0.9;
+    background-color: var(--close);
+    color: var(--white);
+    align-self: flex-start;
+}
